js changed to ommit conflicts
