+++ src/js/firebase.js
@@ -1,5 +1,11 @@
 import { initializeApp } from 'firebase/app';
-import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword } from 'firebase/auth';
+import {
+  getAuth,
+  signInWithEmailAndPassword,
+  createUserWithEmailAndPassword,
+  onAuthStateChanged,
+} from 'firebase/auth';
+import { movieTypes } from './genres.js';
 import { getDatabase, ref, set, child, get, update, remove } from 'firebase/database';
 import { Notify } from 'notiflix';
 
@@ -12,7 +18,7 @@ const firebaseConfig = {
   appId: '1:318653212510:web:dc27e0d5cca4f8c4cfd1cf',
   measurementId: 'G-WJNSN4TTJV',
 };
-
+const movieListEl = document.querySelector('.movie-list');
 const app = initializeApp(firebaseConfig);
 const db = getDatabase();
 const auth = getAuth(app);
@@ -22,6 +28,12 @@ let loginPassword = document.getElementById('login-password').value;
 let user;
 let id;
 let clickedMovie;
+
+onAuthStateChanged(auth, currentUser => {
+  if (currentUser) {
+    user = currentUser;
+  }
+});
 
 //**************** */
 
@@ -51,12 +63,12 @@ document.getElementById('register-btn').addEventListener('click', function () {
   createUserWithEmailAndPassword(auth, registerEmail, registerPassword)
     .then(userCredential => {
       user = userCredential.user;
-      //document.getElementById('register-div').style.display = 'inline';
+      document.getElementById('register-div').style.display = 'inline';
       Notify.success('Succesfully registered! Now log in');
     })
     .catch(error => {
       const errorMessage = error.message;
-      //document.getElementById('register-div').style.display = 'inline';
+      document.getElementById('register-div').style.display = 'inline';
       Notify.failure(`${errorMessage}`);
     });
 });
@@ -98,6 +110,7 @@ function addToWatchedOrQueue(
           overview: about,
           release_date: releaseDate,
           poster_path: posterPath,
+          id: UniqueFilmId,
         };
         update(ref(db), updates);
       } else {
@@ -112,6 +125,7 @@ function addToWatchedOrQueue(
           overview: about,
           release_date: releaseDate,
           poster_path: posterPath,
+          id: UniqueFilmId,
         });
       }
       Notify.success(`added to ${libraryPlace} list`);
@@ -146,7 +160,7 @@ document.querySelector('.modal__button--watched').addEventListener('click', () =
     const rating = document.querySelector('.modal__rating').textContent;
     const popularity = arrayData[1];
     const fullTitle = arrayData[2];
-    const genres = arrayData[3];
+    const genres = document.querySelector('.modal__backdrop').dataset.genres.split(',');
     const about = document.querySelector('.modal__descripton').textContent;
     const releaseDate = clickedMovie.release_date;
     const posterPath = clickedMovie.poster_path;
@@ -186,7 +200,7 @@ document.querySelector('.modal__button--queue').addEventListener('click', () => 
     const rating = document.querySelector('.modal__rating').textContent;
     const popularity = arrayData[1];
     const fullTitle = arrayData[2];
-    const genres = arrayData[3];
+    const genres = document.querySelector('.modal__backdrop').dataset.genres.split(',');
     const about = document.querySelector('.modal__descripton').textContent;
     const releaseDate = clickedMovie.release_date;
     const posterPath = clickedMovie.poster_path;
@@ -209,3 +223,84 @@ document.querySelector('.modal__button--queue').addEventListener('click', () => 
     Notify.failure('No user is signed in.');
   }
 });
+
+const drawMovies = (movies, collection) => {
+  let markup = '';
+  let id = 0;
+  movies.forEach(movie => {
+    let posterUrl = movie.poster_path
+      ? `https://image.tmdb.org/t/p/w500${movie.poster_path}`
+      : `https://www.csaff.org/wp-content/uploads/csaff-no-poster.jpg`;
+    let posterUrlRetina = movie.poster_path
+      ? `https://image.tmdb.org/t/p/w780${movie.poster_path}`
+      : `https://www.csaff.org/wp-content/uploads/csaff-no-poster.jpg`;
+    markup += `
+    <div class="movie-card">
+    <div class="movie-card__poster-container">
+    <img class="movie-card__poster" id="poster_path" data-order=${id++} data-collection=${collection}
+    src="${posterUrl}"
+    srcset="${posterUrl} 1x, ${posterUrlRetina} 2x"
+    alt=""
+    />
+    </div>
+    <div class="movie-card__figcaption">
+        <p class="movie-card__title" id="title">${movie.title.toUpperCase()}</p>
+        <span class="movie-card__genre" id="genre_ids">${movieTypes(movie.genre_ids)} |</span>
+        <span class="movie-card__release-date" id="release_date"> ${movie.release_date.slice(
+          0,
+          4,
+        )}</span>
+        <!-- in JS need to add a script for changing visibility on Homepage: document.querySelector('.movie-card__rating').classList.add('is-hidden') -->
+        <span class="movie-card__rating" id="vote_average">${movie.vote_average}</span>
+    </div>
+</div>`;
+  });
+  return markup;
+};
+
+const loadMovies = markup => {
+  movieListEl.innerHTML = '';
+  movieListEl.innerHTML = markup;
+};
+
+//po kliknięciu Watched przekazuje ścieżkę do Firebase/Watched do funkcji
+document.querySelector('.button__status').addEventListener('click', () => {
+  watchedOrQueue = 'watched';
+  passPathToRenderMoviesFrom(watchedOrQueue);
+});
+
+//po kliknięciu Queue przekazuje ścieżkę do Firebase/queue do funkcji
+document.querySelector('.button__status').nextElementSibling.addEventListener('click', () => {
+  watchedOrQueue = 'queue';
+  passPathToRenderMoviesFrom(watchedOrQueue);
+});
+
+//Przekazuje odpowiednią ścieżkę z obiektami do funkcji renderującej filmy
+function passPathToRenderMoviesFrom(watchedOrQueue) {
+  if (user) {
+    let uid = user.uid;
+    get(child(dbRef, uid + '/' + watchedOrQueue)).then(snapshot => {
+      moviesAddedToWatch = snapshot.val();
+      const arrayOfVideoData = Object.values(moviesAddedToWatch);
+      const arrayOfVideoIds = Object.keys(moviesAddedToWatch);
+      console.log(
+        'Tu przechowywane są tablice z filmami i ich ID : ',
+        arrayOfVideoData,
+        arrayOfVideoIds,
+      );
+      // localStorage.setItem(watchedOrQueue, JSON.stringify(arrayOfVideoData));
+      loadMovies(drawMovies(arrayOfVideoData, watchedOrQueue));
+    });
+  } else {
+    Notify.failure('Sign in first');
+  }
+}
+
+//Usuwanie obiektu z Watched lub Queue po właściwości .UniqueId
+function deleteVideoFromLibrary(dbRef, userId, watchedOrQueue, UniqueFilmId) {
+  remove(child(dbRef, userId + '/' + `${watchedOrQueue}` + '/' + `${UniqueFilmId}`))
+    .then(Notify.success(`Removed ${UniqueFilmId} from ${watchedOrQueue} list`))
+    .catch(function (error) {
+      Notify.failure('Wystąpił błąd podczas usuwania obiektu:', error);
+    });
+}
